#29 Redux auth - Moving service's files into it's own directory
