+++ src/screens/HomeScreen.js
@@ -1,7 +1,7 @@
 import React from "react";
 import HeroUnit from "../components/HeroUnit";
 import HeroBanner from "../components/HeroBanner";
-import AuthService from "../components/AuthService";
+import AuthService from "../services/AuthService";
 import {FormattedMessage} from 'react-intl';
 
 const Auth = new AuthService();
@@ -26,8 +26,8 @@ class HomeScreen extends React.Component {
 
     return (
       <div>
-        <HeroBanner title={<FormattedMessage id='chotavern' defaultMessage="Cho's Tavern" />} 
-        description={<FormattedMessage id='homescreen.welcome' defaultMessage="Let's choose the next achievement you want to do and find new friends" />} 
+        <HeroBanner title={<FormattedMessage id='chotavern' defaultMessage="Cho's Tavern" />}
+        description={<FormattedMessage id='homescreen.welcome' defaultMessage="Let's choose the next achievement you want to do and find new friends" />}
         hasButton={this.state.hasButton}/>
         <HeroUnit title={<FormattedMessage id='homescreen' defaultMessage='Home' />}/>
       </div>
